refactor: extract inline tests from gym/ modules
